#132 tiptap-editor.js 모놀리스를 ES 모듈로 분할
2113줄 단일 파일을 nodes/*.js(커스텀 노드), extensions/*.js,
ui/*.js, upload.js, images.js, interop.js 등 tiptap/ 하위 모듈로
분리한다. 번들러/빌드 스텝 없이 순수 ES 모듈 import만 사용하며,
tiptap-editor.js는 index.html이 참조하는 진입점으로 남아
interop.js를 import해 window.tiptapInterop 계약을 그대로 유지한다.
코드는 모두 그대로 옮기고 모듈 간 공유 심볼만 import/export로 연결
했으므로 에디터 동작·직렬화 포맷·인터롭 시그니처 변화는 없다.
