+++ Vanadium.Note.Web/wwwroot/js/tiptap/extensions/code-block.js
@@ -1,0 +1,20 @@
+import { mergeAttributes } from 'https://esm.sh/@tiptap/core@2'
+import { createLowlight, common } from 'https://esm.sh/lowlight'
+import CodeBlockLowlight from 'https://esm.sh/@tiptap/extension-code-block-lowlight@2'
+
+// ── Code block with lowlight syntax highlighting ─────────────────────────────
+
+export const lowlight = createLowlight(common)
+
+export const CodeBlock = CodeBlockLowlight.extend({
+    renderHTML({ node, HTMLAttributes }) {
+        const lang = node.attrs.language;
+        return [
+            'pre',
+            mergeAttributes(this.options.HTMLAttributes, HTMLAttributes, {
+                'data-language': lang && lang !== 'plaintext' ? lang : null,
+            }),
+            ['code', { class: lang ? `language-${lang}` : null }, 0],
+        ];
+    },
+});
